FEAT: Visual updates, weather card
- Visual layout (more centered & responsive)
- Added weather card
- Server-down svg
- Added rating to sport card

FIXES: #4, #5
