+++ frontend/src/app/app.component.html
@@ -1,14 +1,19 @@
 <!-- Toolbar -->
 <app-toolbar></app-toolbar>
 
-<div class="content" role="main">
-
-  <!-- Greeting -->
+<!-- Greeting -->
+<div class="greeting-card">
   <app-greeting></app-greeting>
+</div>
 
-  <!-- Sport -->
+<!-- Sport -->
+<div class="sport-cards">
   <app-sport></app-sport>
+</div>
 
+<!-- Weather -->
+<div class="weather-card">
+  <app-weather></app-weather>
 </div>
 
 <!-- Footer -->
