+++ frontend/src/app/toolbar/toolbar.component.scss
@@ -7,7 +7,6 @@
   top: 0;
   left: 0;
   right: 0;
-  height: 60px;
   display: flex;
   align-items: center;
   background-color: #762bd8;
